requirements.txt
tensorflow
numpy
scikit-learn
gdown
flake8-qgis
PyQt5-stubs
qgis-stubs
pb_tool